+++ .pre-commit-config.yaml
@@ -44,6 +44,6 @@ repos:
     hooks:
       - id: yamllint
   - repo: https://github.com/domluna/JuliaFormatter.jl
-    rev: v2.1.1
+    rev: v2.1.2
     hooks:
       - id: julia-formatter
