+++ src/Rule/cube.jl
@@ -1,599 +1,4 @@
 """
-Gauss-Kronrod with 7 nodes.
-"""
-const SEGMENT_GK7 = TabulatedEmbeddedCubature{Segment}(;
-    description="Gauss-Kronrod with 7 nodes (SEGMENT_GK7)",
-    reference="https://github.com/JuliaMath/QuadGK.jl",
-    precision=35,
-    nodes=[
-        ["5.000000000000000000000000000000000000e-01"],
-        ["8.872983346207416885179265399782399611e-01"],
-        ["1.127016653792583114820734600217600389e-01"],
-        ["7.171218746734012790010357514223139086e-01"],
-        ["2.828781253265987209989642485776860914e-01"],
-        ["9.802456343540101417117535463145399813e-01"],
-        ["1.975436564598985828824645368546001867e-02"],
-    ],
-    weights_high=[
-        "2.254582693292370711725550435227854583e-01",
-        "1.342440449341667203642846403333548124e-01",
-        "1.342440449341667203642846403333548124e-01",
-        "2.006987073879811114525259093092159394e-01",
-        "2.006987073879811114525259093092159394e-01",
-        "5.232811301323363259691192859603651912e-02",
-        "5.232811301323363259691192859603651912e-02",
-    ],
-    order_high=11,
-    weights_low=[
-        "4.444444444444444444444444444444444444e-01",
-        "2.777777777777777777777777777777777778e-01",
-        "2.777777777777777777777777777777777778e-01",
-    ],
-    order_low=5,
-)
-
-"""
-Gauss-Kronrod with 15 nodes.
-"""
-const SEGMENT_GK15 = TabulatedEmbeddedCubature{Segment}(;
-    description="Gauss-Kronrod with 15 nodes (SEGMENT_GK15)",
-    reference="https://github.com/JuliaMath/QuadGK.jl",
-    precision=35,
-    nodes=[
-        ["5.000000000000000000000000000000000000e-01"],
-        ["7.029225756886985834533032060384807317e-01"],
-        ["2.970774243113014165466967939615192683e-01"],
-        ["8.707655927996972199319323866403942035e-01"],
-        ["1.292344072003027800680676133596057965e-01"],
-        ["9.745539561713792622630948420239256312e-01"],
-        ["2.544604382862073773690515797607436880e-02"],
-        ["6.038924775039492338003447018866224567e-01"],
-        ["3.961075224960507661996552981133775433e-01"],
-        ["7.930436177338455651470724191293647992e-01"],
-        ["2.069563822661544348529275808706352008e-01"],
-        ["9.324322116798845363948563943204631006e-01"],
-        ["6.756778832011546360514360567953689939e-02"],
-        ["9.957276855604063196034273487631642583e-01"],
-        ["4.272314439593680396572651236835741679e-03"],
-    ],
-    weights_high=[
-        "1.047410705423639140064995874458571318e-01",
-        "9.517528903239270495662820121050684141e-02",
-        "9.517528903239270495662820121050684141e-02",
-        "7.032662985776295937259479525511896020e-02",
-        "7.032662985776295937259479525511896020e-02",
-        "3.154604631498927664535033159460214333e-02",
-        "3.154604631498927664535033159460214333e-02",
-        "1.022164700376494462070809996173245424e-01",
-        "1.022164700376494462070809996173245424e-01",
-        "8.450236331963395141329171329927514205e-02",
-        "8.450236331963395141329171329927514205e-02",
-        "5.239500516112509191993816127075900872e-02",
-        "5.239500516112509191993816127075900872e-02",
-        "1.146766100526461248186600402948479600e-02",
-        "1.146766100526461248186600402948479600e-02",
-    ],
-    order_high=23,
-    weights_low=[
-        "2.089795918367346938775510204081632653e-01",
-        "1.909150252525594724751848877444875669e-01",
-        "1.909150252525594724751848877444875669e-01",
-        "1.398526957446383339507338857118897912e-01",
-        "1.398526957446383339507338857118897912e-01",
-        "6.474248308443484663530571633954100916e-02",
-        "6.474248308443484663530571633954100916e-02",
-    ],
-    order_low=13,
-)
-
-"""
-Gauss-Kronrod with 31 nodes.
-"""
-const SEGMENT_GK31 = TabulatedEmbeddedCubature{Segment}(;
-    description="Gauss-Kronrod with 31 nodes",
-    reference="https://github.com/JuliaMath/QuadGK.jl",
-    precision=35,
-    nodes=[
-        ["5.000000000000000000000000000000000000e-01"],
-        ["6.005970469987172611503141516972981039e-01"],
-        ["3.994029530012827388496858483027018961e-01"],
-        ["6.970756735387816849486036854905227342e-01"],
-        ["3.029243264612183150513963145094772658e-01"],
-        ["7.854860863042694237686133686269553206e-01"],
-        ["2.145139136957305762313866313730446794e-01"],
-        ["8.622088656800850237080930273069690048e-01"],
-        ["1.377911343199149762919069726930309952e-01"],
-        ["9.241032917052136081003241603871084257e-01"],
-        ["7.589670829478639189967583961289157432e-02"],
-        ["9.686366962003529521538794738551047356e-01"],
-        ["3.136330379964704784612052614489526438e-02"],
-        ["9.939962590102427142447828592933062906e-01"],
-        ["6.003740989757285755217140706693709427e-03"],
-        ["5.505710334593587495135371157236961694e-01"],
-        ["4.494289665406412504864628842763038306e-01"],
-        ["6.495900035765844060833900121331944813e-01"],
-        ["3.504099964234155939166099878668055187e-01"],
-        ["7.425409318201198403468278701161753064e-01"],
-        ["2.574590681798801596531721298838246936e-01"],
-        ["8.254983706487084852668679476566373463e-01"],
-        ["1.745016293512915147331320523433626537e-01"],
-        ["8.952092507212329664838246474089736734e-01"],
-        ["1.047907492787670335161753525910263266e-01"],
-        ["9.486322661720409504412548282272479414e-01"],
-        ["5.136773382795904955874517177275205858e-02"],
-        ["9.838695378395695671286739893921686126e-01"],
-        ["1.613046216043043287132601060783138736e-02"],
-        ["9.990011493466985301425864200761356045e-01"],
-        ["9.988506533014698574135799238643954633e-04"],
-    ],
-    weights_high=[
-        "5.066500350739577450868739638374627339e-02",
-        "4.958679936089597966619658674230156553e-02",
-        "4.958679936089597966619658674230156553e-02",
-        "4.656329908541266061274343637367285928e-02",
-        "4.656329908541266061274343637367285928e-02",
-        "4.154025141156651051914462364305189480e-02",
-        "4.154025141156651051914462364305189480e-02",
-        "3.492706065936412935476003854957373789e-02",
-        "3.492706065936412935476003854957373789e-02",
-        "2.674076234546404363267157361971514839e-02",
-        "2.674076234546404363267157361971514839e-02",
-        "1.767318039568792311101897423918002406e-02",
-        "1.767318039568792311101897423918002406e-02",
-        "7.503973664658061269187381537903634047e-03",
-        "7.503973664658061269187381537903634047e-03",
-        "5.038492276193779752247333130878486096e-02",
-        "5.038492276193779752247333130878486096e-02",
-        "4.832136349181183925258995381379466757e-02",
-        "4.832136349181183925258995381379466757e-02",
-        "4.428222152810588532363772184688715161e-02",
-        "4.428222152810588532363772184688715161e-02",
-        "3.842484037886018944721638874132950336e-02",
-        "3.842484037886018944721638874132950336e-02",
-        "3.100478390033532014256961548040146610e-02",
-        "3.100478390033532014256961548040146610e-02",
-        "2.229487566238243830411364968663984511e-02",
-        "2.229487566238243830411364968663984511e-02",
-        "1.273042366335766009343700050982667970e-02",
-        "1.273042366335766009343700050982667970e-02",
-        "2.688739936461674493896025715063824909e-03",
-        "2.688739936461674493896025715063824909e-03",
-    ],
-    order_high=47,
-    weights_low=[
-        "1.012891209627806364403100999837596574e-01",
-        "9.921574266355578822805916322191966241e-02",
-        "9.921574266355578822805916322191966241e-02",
-        "9.308050000778110551340028093321141225e-02",
-        "9.308050000778110551340028093321141225e-02",
-        "8.313460290849696677660043024060440557e-02",
-        "8.313460290849696677660043024060440557e-02",
-        "6.978533896307715722390239725551416126e-02",
-        "6.978533896307715722390239725551416126e-02",
-        "5.357961023358596750593477334293465171e-02",
-        "5.357961023358596750593477334293465171e-02",
-        "3.518302374405406235463370822533366923e-02",
-        "3.518302374405406235463370822533366923e-02",
-        "1.537662099805863417731419678860220886e-02",
-        "1.537662099805863417731419678860220886e-02",
-    ],
-    order_low=29,
-)
-
-"""
-Genz-Malik with 17 nodes.
-"""
-const SQUARE_GM17 = TabulatedEmbeddedCubature{Rectangle}(;
-    description="Genz-Malik with 17 nodes (SQUARE_GM17)",
-    reference="https://doi.org/10.1016/0771-050X(80)90039-X",
-    precision=35,
-    nodes=[
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "6.792842914001590459953225769539687477e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "3.207157085998409540046774230460312523e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "6.792842914001590459953225769539687477e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "3.207157085998409540046774230460312523e-01",
-        ],
-        [
-            "9.743416490252568997998340316649077801e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "2.565835097474310020016596833509221994e-02",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "9.743416490252568997998340316649077801e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "2.565835097474310020016596833509221994e-02",
-        ],
-        [
-            "9.743416490252568997998340316649077801e-01",
-            "9.743416490252568997998340316649077801e-01",
-        ],
-        [
-            "2.565835097474310020016596833509221994e-02",
-            "9.743416490252568997998340316649077801e-01",
-        ],
-        [
-            "9.743416490252568997998340316649077801e-01",
-            "2.565835097474310020016596833509221994e-02",
-        ],
-        [
-            "2.565835097474310020016596833509221994e-02",
-            "2.565835097474310020016596833509221994e-02",
-        ],
-        [
-            "8.441236008058426488608143671468117626e-01",
-            "8.441236008058426488608143671468117626e-01",
-        ],
-        [
-            "1.558763991941573511391856328531882374e-01",
-            "8.441236008058426488608143671468117626e-01",
-        ],
-        [
-            "8.441236008058426488608143671468117626e-01",
-            "1.558763991941573511391856328531882374e-01",
-        ],
-        [
-            "1.558763991941573511391856328531882374e-01",
-            "1.558763991941573511391856328531882374e-01",
-        ],
-    ],
-    weights_high=[
-        "-1.938728852309099222679469593049839963e-01",
-        "1.493674744703551287913427831123304374e-01",
-        "1.493674744703551287913427831123304374e-01",
-        "1.493674744703551287913427831123304374e-01",
-        "1.493674744703551287913427831123304374e-01",
-        "5.182136869379667733577198597774729462e-02",
-        "5.182136869379667733577198597774729462e-02",
-        "5.182136869379667733577198597774729462e-02",
-        "5.182136869379667733577198597774729462e-02",
-        "1.016105268505817202662195803485241071e-02",
-        "1.016105268505817202662195803485241071e-02",
-        "1.016105268505817202662195803485241071e-02",
-        "1.016105268505817202662195803485241071e-02",
-        "8.711832545851750241325001270131585632e-02",
-        "8.711832545851750241325001270131585632e-02",
-        "8.711832545851750241325001270131585632e-02",
-        "8.711832545851750241325001270131585632e-02",
-    ],
-    order_high=7,
-    weights_low=[
-        "-1.331961591220850480109739368998628258e+00",
-        "5.041152263374485596707818930041152263e-01",
-        "5.041152263374485596707818930041152263e-01",
-        "5.041152263374485596707818930041152263e-01",
-        "5.041152263374485596707818930041152263e-01",
-        "4.458161865569272976680384087791495199e-02",
-        "4.458161865569272976680384087791495199e-02",
-        "4.458161865569272976680384087791495199e-02",
-        "4.458161865569272976680384087791495199e-02",
-        "3.429355281207133058984910836762688615e-02",
-        "3.429355281207133058984910836762688615e-02",
-        "3.429355281207133058984910836762688615e-02",
-        "3.429355281207133058984910836762688615e-02",
-    ],
-    order_low=5,
-)
-
-"""
-Cools-Haegemans with 21 nodes.
-"""
-const SQUARE_CH21 = TabulatedEmbeddedCubature{Rectangle}(;
-    description="Cools-Haegemans with 21 nodes (SQUARE_CH21)",
-    reference="https://link.springer.com/article/10.1007/BF01389339",
-    precision=35,
-    nodes=[
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "9.530899229693320005789205883895798105e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "9.530899229693320005789205883895798105e-01",
-        ],
-        [
-            "4.691007703066799942107941161042018947e-02",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "4.691007703066799942107941161042018947e-02",
-        ],
-        [
-            "7.692346550528415200375282027103137048e-01",
-            "7.692346550528415200375282027103137048e-01",
-        ],
-        [
-            "2.307653449471584799624717972896862952e-01",
-            "7.692346550528415200375282027103137048e-01",
-        ],
-        [
-            "7.692346550528415200375282027103137048e-01",
-            "2.307653449471584799624717972896862952e-01",
-        ],
-        [
-            "2.307653449471584799624717972896862952e-01",
-            "2.307653449471584799624717972896862952e-01",
-        ],
-        [
-            "9.530899229693319571124065987079251159e-01",
-            "9.530899229693319571124065987079251159e-01",
-        ],
-        [
-            "4.691007703066804288759340129207488409e-02",
-            "9.530899229693319571124065987079251159e-01",
-        ],
-        [
-            "9.530899229693319571124065987079251159e-01",
-            "4.691007703066804288759340129207488409e-02",
-        ],
-        [
-            "4.691007703066804288759340129207488409e-02",
-            "4.691007703066804288759340129207488409e-02",
-        ],
-        [
-            "9.530899229693320081691800769275475942e-01",
-            "7.692346550528415524068181972304942518e-01",
-        ],
-        [
-            "7.692346550528415524068181972304942518e-01",
-            "9.530899229693320081691800769275475942e-01",
-        ],
-        [
-            "4.691007703066799183081992307245240580e-02",
-            "7.692346550528415524068181972304942518e-01",
-        ],
-        [
-            "7.692346550528415524068181972304942518e-01",
-            "4.691007703066799183081992307245240580e-02",
-        ],
-        [
-            "9.530899229693320081691800769275475942e-01",
-            "2.307653449471584475931818027695057482e-01",
-        ],
-        [
-            "2.307653449471584475931818027695057482e-01",
-            "9.530899229693320081691800769275475942e-01",
-        ],
-        [
-            "4.691007703066799183081992307245240580e-02",
-            "2.307653449471584475931818027695057482e-01",
-        ],
-        [
-            "2.307653449471584475931818027695057482e-01",
-            "4.691007703066799183081992307245240580e-02",
-        ],
-    ],
-    weights_high=[
-        "1.689802305149250673159321255328970433e-01",
-        "5.773210696475967157451779245157955517e-02",
-        "5.773210696475967157451779245157955517e-02",
-        "5.773210696475967157451779245157955517e-02",
-        "5.773210696475967157451779245157955517e-02",
-        "1.098847683324169696623710997801633006e-01",
-        "1.098847683324169696623710997801633006e-01",
-        "1.098847683324169696623710997801633006e-01",
-        "1.098847683324169696623710997801633006e-01",
-        "2.059326848924280095905724771232335608e-02",
-        "2.059326848924280095905724771232335608e-02",
-        "2.059326848924280095905724771232335608e-02",
-        "2.059326848924280095905724771232335608e-02",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-        "9.772399292424645487535414336354763648e-03",
-    ],
-    order_high=7,
-    weights_low=[
-        "1.526218418361306486442884511393218447e-01",
-        "6.591130130415688306506220703853742741e-02",
-        "6.591130130415688306506220703853742741e-02",
-        "6.591130130415688306506220703853742741e-02",
-        "6.591130130415688306506220703853742741e-02",
-        "1.196571676248416100102967077035190968e-01",
-        "1.196571676248416100102967077035190968e-01",
-        "1.196571676248416100102967077035190968e-01",
-        "1.196571676248416100102967077035190968e-01",
-        "2.627607061196884476356897247311301463e-02",
-        "2.627607061196884476356897247311301463e-02",
-        "2.627607061196884476356897247311301463e-02",
-        "2.627607061196884476356897247311301463e-02",
-    ],
-    order_low=5,
-)
-
-"""
-Cools-Haegemans with 25 nodes.
-"""
-const SQUARE_CH25 = TabulatedEmbeddedCubature{Rectangle}(;
-    description="Cools-Haegemans with 25 nodes (SQUARE_CH25)",
-    reference="https://link.springer.com/article/10.1007/BF01389339",
-    precision=35,
-    nodes=[
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "9.530899230491018626945996533246312660e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "9.530899230491018626945996533246312660e-01",
-        ],
-        [
-            "4.691007695089813730540034667536873396e-02",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "4.691007695089813730540034667536873396e-02",
-        ],
-        [
-            "7.692346551549900625567506592830037596e-01",
-            "7.692346551549900625567506592830037596e-01",
-        ],
-        [
-            "2.307653448450099374432493407169962404e-01",
-            "7.692346551549900625567506592830037596e-01",
-        ],
-        [
-            "7.692346551549900625567506592830037596e-01",
-            "2.307653448450099374432493407169962404e-01",
-        ],
-        [
-            "2.307653448450099374432493407169962404e-01",
-            "2.307653448450099374432493407169962404e-01",
-        ],
-        [
-            "9.530899229853897579996255404928620237e-01",
-            "9.530899229853897579996255404928620237e-01",
-        ],
-        [
-            "4.691007701461024200037445950713797632e-02",
-            "9.530899229853897579996255404928620237e-01",
-        ],
-        [
-            "9.530899229853897579996255404928620237e-01",
-            "4.691007701461024200037445950713797632e-02",
-        ],
-        [
-            "4.691007701461024200037445950713797632e-02",
-            "4.691007701461024200037445950713797632e-02",
-        ],
-        [
-            "9.530899229139767442924391519556961148e-01",
-            "7.692346549247062656221145758591156811e-01",
-        ],
-        [
-            "7.692346549247062656221145758591156811e-01",
-            "9.530899229139767442924391519556961148e-01",
-        ],
-        [
-            "4.691007708602325570756084804430388521e-02",
-            "7.692346549247062656221145758591156811e-01",
-        ],
-        [
-            "7.692346549247062656221145758591156811e-01",
-            "4.691007708602325570756084804430388521e-02",
-        ],
-        [
-            "9.530899229139767442924391519556961148e-01",
-            "2.307653450752937343778854241408843189e-01",
-        ],
-        [
-            "2.307653450752937343778854241408843189e-01",
-            "9.530899229139767442924391519556961148e-01",
-        ],
-        [
-            "4.691007708602325570756084804430388521e-02",
-            "2.307653450752937343778854241408843189e-01",
-        ],
-        [
-            "2.307653450752937343778854241408843189e-01",
-            "4.691007708602325570756084804430388521e-02",
-        ],
-        [
-            "7.692346549876881686600646748390376768e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "7.692346549876881686600646748390376768e-01",
-        ],
-        [
-            "2.307653450123118313399353251609623232e-01",
-            "5.000000000000000000000000000000000000e-01",
-        ],
-        [
-            "5.000000000000000000000000000000000000e-01",
-            "2.307653450123118313399353251609623232e-01",
-        ],
-    ],
-    weights_high=[
-        "8.090864180310654470524179109436875230e-02",
-        "3.369626803446566471100225222749401905e-02",
-        "3.369626803446566471100225222749401905e-02",
-        "3.369626803446566471100225222749401905e-02",
-        "3.369626803446566471100225222749401905e-02",
-        "5.727135095586947958087052565817859454e-02",
-        "5.727135095586947958087052565817859454e-02",
-        "5.727135095586947958087052565817859454e-02",
-        "5.727135095586947958087052565817859454e-02",
-        "1.403358722094394462690571699238021305e-02",
-        "1.403358722094394462690571699238021305e-02",
-        "1.403358722094394462690571699238021305e-02",
-        "1.403358722094394462690571699238021305e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "2.835000002587049491397105606308867037e-02",
-        "6.807163328620328507696894522217764455e-02",
-        "6.807163328620328507696894522217764455e-02",
-        "6.807163328620328507696894522217764455e-02",
-        "6.807163328620328507696894522217764455e-02",
-    ],
-    order_high=9,
-    weights_low=[
-        "1.689802306319903682683290932522631273e-01",
-        "5.773210696822923076352990074033502715e-02",
-        "5.773210696822923076352990074033502715e-02",
-        "5.773210696822923076352990074033502715e-02",
-        "5.773210696822923076352990074033502715e-02",
-        "1.098847683946998119578127419601697553e-01",
-        "1.098847683946998119578127419601697553e-01",
-        "1.098847683946998119578127419601697553e-01",
-        "1.098847683946998119578127419601697553e-01",
-        "2.059326850521098279376569809479221704e-02",
-        "2.059326850521098279376569809479221704e-02",
-        "2.059326850521098279376569809479221704e-02",
-        "2.059326850521098279376569809479221704e-02",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-        "9.772399236931191208904692945818609348e-03",
-    ],
-    order_low=7,
-)
-
-"""
 Genz-Malik with 33 nodes.
 """
 const CUBE_GM33 = TabulatedEmbeddedCubature{Cuboid}(;
