+++ src/Rule/segment.jl
@@ -1,0 +1,181 @@
+"""
+Gauss-Kronrod with 7 nodes.
+"""
+const SEGMENT_GK7 = TabulatedEmbeddedCubature{Segment}(;
+    description="Gauss-Kronrod with 7 nodes (SEGMENT_GK7)",
+    reference="https://github.com/JuliaMath/QuadGK.jl",
+    precision=35,
+    nodes=[
+        ["5.000000000000000000000000000000000000e-01"],
+        ["8.872983346207416885179265399782399611e-01"],
+        ["1.127016653792583114820734600217600389e-01"],
+        ["7.171218746734012790010357514223139086e-01"],
+        ["2.828781253265987209989642485776860914e-01"],
+        ["9.802456343540101417117535463145399813e-01"],
+        ["1.975436564598985828824645368546001867e-02"],
+    ],
+    weights_high=[
+        "2.254582693292370711725550435227854583e-01",
+        "1.342440449341667203642846403333548124e-01",
+        "1.342440449341667203642846403333548124e-01",
+        "2.006987073879811114525259093092159394e-01",
+        "2.006987073879811114525259093092159394e-01",
+        "5.232811301323363259691192859603651912e-02",
+        "5.232811301323363259691192859603651912e-02",
+    ],
+    order_high=11,
+    weights_low=[
+        "4.444444444444444444444444444444444444e-01",
+        "2.777777777777777777777777777777777778e-01",
+        "2.777777777777777777777777777777777778e-01",
+    ],
+    order_low=5,
+)
+
+"""
+Gauss-Kronrod with 15 nodes.
+"""
+const SEGMENT_GK15 = TabulatedEmbeddedCubature{Segment}(;
+    description="Gauss-Kronrod with 15 nodes (SEGMENT_GK15)",
+    reference="https://github.com/JuliaMath/QuadGK.jl",
+    precision=35,
+    nodes=[
+        ["5.000000000000000000000000000000000000e-01"],
+        ["7.029225756886985834533032060384807317e-01"],
+        ["2.970774243113014165466967939615192683e-01"],
+        ["8.707655927996972199319323866403942035e-01"],
+        ["1.292344072003027800680676133596057965e-01"],
+        ["9.745539561713792622630948420239256312e-01"],
+        ["2.544604382862073773690515797607436880e-02"],
+        ["6.038924775039492338003447018866224567e-01"],
+        ["3.961075224960507661996552981133775433e-01"],
+        ["7.930436177338455651470724191293647992e-01"],
+        ["2.069563822661544348529275808706352008e-01"],
+        ["9.324322116798845363948563943204631006e-01"],
+        ["6.756778832011546360514360567953689939e-02"],
+        ["9.957276855604063196034273487631642583e-01"],
+        ["4.272314439593680396572651236835741679e-03"],
+    ],
+    weights_high=[
+        "1.047410705423639140064995874458571318e-01",
+        "9.517528903239270495662820121050684141e-02",
+        "9.517528903239270495662820121050684141e-02",
+        "7.032662985776295937259479525511896020e-02",
+        "7.032662985776295937259479525511896020e-02",
+        "3.154604631498927664535033159460214333e-02",
+        "3.154604631498927664535033159460214333e-02",
+        "1.022164700376494462070809996173245424e-01",
+        "1.022164700376494462070809996173245424e-01",
+        "8.450236331963395141329171329927514205e-02",
+        "8.450236331963395141329171329927514205e-02",
+        "5.239500516112509191993816127075900872e-02",
+        "5.239500516112509191993816127075900872e-02",
+        "1.146766100526461248186600402948479600e-02",
+        "1.146766100526461248186600402948479600e-02",
+    ],
+    order_high=23,
+    weights_low=[
+        "2.089795918367346938775510204081632653e-01",
+        "1.909150252525594724751848877444875669e-01",
+        "1.909150252525594724751848877444875669e-01",
+        "1.398526957446383339507338857118897912e-01",
+        "1.398526957446383339507338857118897912e-01",
+        "6.474248308443484663530571633954100916e-02",
+        "6.474248308443484663530571633954100916e-02",
+    ],
+    order_low=13,
+)
+
+"""
+Gauss-Kronrod with 31 nodes.
+"""
+const SEGMENT_GK31 = TabulatedEmbeddedCubature{Segment}(;
+    description="Gauss-Kronrod with 31 nodes",
+    reference="https://github.com/JuliaMath/QuadGK.jl",
+    precision=35,
+    nodes=[
+        ["5.000000000000000000000000000000000000e-01"],
+        ["6.005970469987172611503141516972981039e-01"],
+        ["3.994029530012827388496858483027018961e-01"],
+        ["6.970756735387816849486036854905227342e-01"],
+        ["3.029243264612183150513963145094772658e-01"],
+        ["7.854860863042694237686133686269553206e-01"],
+        ["2.145139136957305762313866313730446794e-01"],
+        ["8.622088656800850237080930273069690048e-01"],
+        ["1.377911343199149762919069726930309952e-01"],
+        ["9.241032917052136081003241603871084257e-01"],
+        ["7.589670829478639189967583961289157432e-02"],
+        ["9.686366962003529521538794738551047356e-01"],
+        ["3.136330379964704784612052614489526438e-02"],
+        ["9.939962590102427142447828592933062906e-01"],
+        ["6.003740989757285755217140706693709427e-03"],
+        ["5.505710334593587495135371157236961694e-01"],
+        ["4.494289665406412504864628842763038306e-01"],
+        ["6.495900035765844060833900121331944813e-01"],
+        ["3.504099964234155939166099878668055187e-01"],
+        ["7.425409318201198403468278701161753064e-01"],
+        ["2.574590681798801596531721298838246936e-01"],
+        ["8.254983706487084852668679476566373463e-01"],
+        ["1.745016293512915147331320523433626537e-01"],
+        ["8.952092507212329664838246474089736734e-01"],
+        ["1.047907492787670335161753525910263266e-01"],
+        ["9.486322661720409504412548282272479414e-01"],
+        ["5.136773382795904955874517177275205858e-02"],
+        ["9.838695378395695671286739893921686126e-01"],
+        ["1.613046216043043287132601060783138736e-02"],
+        ["9.990011493466985301425864200761356045e-01"],
+        ["9.988506533014698574135799238643954633e-04"],
+    ],
+    weights_high=[
+        "5.066500350739577450868739638374627339e-02",
+        "4.958679936089597966619658674230156553e-02",
+        "4.958679936089597966619658674230156553e-02",
+        "4.656329908541266061274343637367285928e-02",
+        "4.656329908541266061274343637367285928e-02",
+        "4.154025141156651051914462364305189480e-02",
+        "4.154025141156651051914462364305189480e-02",
+        "3.492706065936412935476003854957373789e-02",
+        "3.492706065936412935476003854957373789e-02",
+        "2.674076234546404363267157361971514839e-02",
+        "2.674076234546404363267157361971514839e-02",
+        "1.767318039568792311101897423918002406e-02",
+        "1.767318039568792311101897423918002406e-02",
+        "7.503973664658061269187381537903634047e-03",
+        "7.503973664658061269187381537903634047e-03",
+        "5.038492276193779752247333130878486096e-02",
+        "5.038492276193779752247333130878486096e-02",
+        "4.832136349181183925258995381379466757e-02",
+        "4.832136349181183925258995381379466757e-02",
+        "4.428222152810588532363772184688715161e-02",
+        "4.428222152810588532363772184688715161e-02",
+        "3.842484037886018944721638874132950336e-02",
+        "3.842484037886018944721638874132950336e-02",
+        "3.100478390033532014256961548040146610e-02",
+        "3.100478390033532014256961548040146610e-02",
+        "2.229487566238243830411364968663984511e-02",
+        "2.229487566238243830411364968663984511e-02",
+        "1.273042366335766009343700050982667970e-02",
+        "1.273042366335766009343700050982667970e-02",
+        "2.688739936461674493896025715063824909e-03",
+        "2.688739936461674493896025715063824909e-03",
+    ],
+    order_high=47,
+    weights_low=[
+        "1.012891209627806364403100999837596574e-01",
+        "9.921574266355578822805916322191966241e-02",
+        "9.921574266355578822805916322191966241e-02",
+        "9.308050000778110551340028093321141225e-02",
+        "9.308050000778110551340028093321141225e-02",
+        "8.313460290849696677660043024060440557e-02",
+        "8.313460290849696677660043024060440557e-02",
+        "6.978533896307715722390239725551416126e-02",
+        "6.978533896307715722390239725551416126e-02",
+        "5.357961023358596750593477334293465171e-02",
+        "5.357961023358596750593477334293465171e-02",
+        "3.518302374405406235463370822533366923e-02",
+        "3.518302374405406235463370822533366923e-02",
+        "1.537662099805863417731419678860220886e-02",
+        "1.537662099805863417731419678860220886e-02",
+    ],
+    order_low=29,
+)
