+++ src/Rule/square.jl
@@ -1,0 +1,412 @@
+"""
+Genz-Malik with 17 nodes.
+"""
+const SQUARE_GM17 = TabulatedEmbeddedCubature{Rectangle}(;
+    description="Genz-Malik with 17 nodes (SQUARE_GM17)",
+    reference="https://doi.org/10.1016/0771-050X(80)90039-X",
+    precision=35,
+    nodes=[
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "6.792842914001590459953225769539687477e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "3.207157085998409540046774230460312523e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "6.792842914001590459953225769539687477e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "3.207157085998409540046774230460312523e-01",
+        ],
+        [
+            "9.743416490252568997998340316649077801e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "2.565835097474310020016596833509221994e-02",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "9.743416490252568997998340316649077801e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "2.565835097474310020016596833509221994e-02",
+        ],
+        [
+            "9.743416490252568997998340316649077801e-01",
+            "9.743416490252568997998340316649077801e-01",
+        ],
+        [
+            "2.565835097474310020016596833509221994e-02",
+            "9.743416490252568997998340316649077801e-01",
+        ],
+        [
+            "9.743416490252568997998340316649077801e-01",
+            "2.565835097474310020016596833509221994e-02",
+        ],
+        [
+            "2.565835097474310020016596833509221994e-02",
+            "2.565835097474310020016596833509221994e-02",
+        ],
+        [
+            "8.441236008058426488608143671468117626e-01",
+            "8.441236008058426488608143671468117626e-01",
+        ],
+        [
+            "1.558763991941573511391856328531882374e-01",
+            "8.441236008058426488608143671468117626e-01",
+        ],
+        [
+            "8.441236008058426488608143671468117626e-01",
+            "1.558763991941573511391856328531882374e-01",
+        ],
+        [
+            "1.558763991941573511391856328531882374e-01",
+            "1.558763991941573511391856328531882374e-01",
+        ],
+    ],
+    weights_high=[
+        "-1.938728852309099222679469593049839963e-01",
+        "1.493674744703551287913427831123304374e-01",
+        "1.493674744703551287913427831123304374e-01",
+        "1.493674744703551287913427831123304374e-01",
+        "1.493674744703551287913427831123304374e-01",
+        "5.182136869379667733577198597774729462e-02",
+        "5.182136869379667733577198597774729462e-02",
+        "5.182136869379667733577198597774729462e-02",
+        "5.182136869379667733577198597774729462e-02",
+        "1.016105268505817202662195803485241071e-02",
+        "1.016105268505817202662195803485241071e-02",
+        "1.016105268505817202662195803485241071e-02",
+        "1.016105268505817202662195803485241071e-02",
+        "8.711832545851750241325001270131585632e-02",
+        "8.711832545851750241325001270131585632e-02",
+        "8.711832545851750241325001270131585632e-02",
+        "8.711832545851750241325001270131585632e-02",
+    ],
+    order_high=7,
+    weights_low=[
+        "-1.331961591220850480109739368998628258e+00",
+        "5.041152263374485596707818930041152263e-01",
+        "5.041152263374485596707818930041152263e-01",
+        "5.041152263374485596707818930041152263e-01",
+        "5.041152263374485596707818930041152263e-01",
+        "4.458161865569272976680384087791495199e-02",
+        "4.458161865569272976680384087791495199e-02",
+        "4.458161865569272976680384087791495199e-02",
+        "4.458161865569272976680384087791495199e-02",
+        "3.429355281207133058984910836762688615e-02",
+        "3.429355281207133058984910836762688615e-02",
+        "3.429355281207133058984910836762688615e-02",
+        "3.429355281207133058984910836762688615e-02",
+    ],
+    order_low=5,
+)
+
+"""
+Cools-Haegemans with 21 nodes.
+"""
+const SQUARE_CH21 = TabulatedEmbeddedCubature{Rectangle}(;
+    description="Cools-Haegemans with 21 nodes (SQUARE_CH21)",
+    reference="https://link.springer.com/article/10.1007/BF01389339",
+    precision=35,
+    nodes=[
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "9.530899229693320005789205883895798105e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "9.530899229693320005789205883895798105e-01",
+        ],
+        [
+            "4.691007703066799942107941161042018947e-02",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "4.691007703066799942107941161042018947e-02",
+        ],
+        [
+            "7.692346550528415200375282027103137048e-01",
+            "7.692346550528415200375282027103137048e-01",
+        ],
+        [
+            "2.307653449471584799624717972896862952e-01",
+            "7.692346550528415200375282027103137048e-01",
+        ],
+        [
+            "7.692346550528415200375282027103137048e-01",
+            "2.307653449471584799624717972896862952e-01",
+        ],
+        [
+            "2.307653449471584799624717972896862952e-01",
+            "2.307653449471584799624717972896862952e-01",
+        ],
+        [
+            "9.530899229693319571124065987079251159e-01",
+            "9.530899229693319571124065987079251159e-01",
+        ],
+        [
+            "4.691007703066804288759340129207488409e-02",
+            "9.530899229693319571124065987079251159e-01",
+        ],
+        [
+            "9.530899229693319571124065987079251159e-01",
+            "4.691007703066804288759340129207488409e-02",
+        ],
+        [
+            "4.691007703066804288759340129207488409e-02",
+            "4.691007703066804288759340129207488409e-02",
+        ],
+        [
+            "9.530899229693320081691800769275475942e-01",
+            "7.692346550528415524068181972304942518e-01",
+        ],
+        [
+            "7.692346550528415524068181972304942518e-01",
+            "9.530899229693320081691800769275475942e-01",
+        ],
+        [
+            "4.691007703066799183081992307245240580e-02",
+            "7.692346550528415524068181972304942518e-01",
+        ],
+        [
+            "7.692346550528415524068181972304942518e-01",
+            "4.691007703066799183081992307245240580e-02",
+        ],
+        [
+            "9.530899229693320081691800769275475942e-01",
+            "2.307653449471584475931818027695057482e-01",
+        ],
+        [
+            "2.307653449471584475931818027695057482e-01",
+            "9.530899229693320081691800769275475942e-01",
+        ],
+        [
+            "4.691007703066799183081992307245240580e-02",
+            "2.307653449471584475931818027695057482e-01",
+        ],
+        [
+            "2.307653449471584475931818027695057482e-01",
+            "4.691007703066799183081992307245240580e-02",
+        ],
+    ],
+    weights_high=[
+        "1.689802305149250673159321255328970433e-01",
+        "5.773210696475967157451779245157955517e-02",
+        "5.773210696475967157451779245157955517e-02",
+        "5.773210696475967157451779245157955517e-02",
+        "5.773210696475967157451779245157955517e-02",
+        "1.098847683324169696623710997801633006e-01",
+        "1.098847683324169696623710997801633006e-01",
+        "1.098847683324169696623710997801633006e-01",
+        "1.098847683324169696623710997801633006e-01",
+        "2.059326848924280095905724771232335608e-02",
+        "2.059326848924280095905724771232335608e-02",
+        "2.059326848924280095905724771232335608e-02",
+        "2.059326848924280095905724771232335608e-02",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+        "9.772399292424645487535414336354763648e-03",
+    ],
+    order_high=7,
+    weights_low=[
+        "1.526218418361306486442884511393218447e-01",
+        "6.591130130415688306506220703853742741e-02",
+        "6.591130130415688306506220703853742741e-02",
+        "6.591130130415688306506220703853742741e-02",
+        "6.591130130415688306506220703853742741e-02",
+        "1.196571676248416100102967077035190968e-01",
+        "1.196571676248416100102967077035190968e-01",
+        "1.196571676248416100102967077035190968e-01",
+        "1.196571676248416100102967077035190968e-01",
+        "2.627607061196884476356897247311301463e-02",
+        "2.627607061196884476356897247311301463e-02",
+        "2.627607061196884476356897247311301463e-02",
+        "2.627607061196884476356897247311301463e-02",
+    ],
+    order_low=5,
+)
+
+"""
+Cools-Haegemans with 25 nodes.
+"""
+const SQUARE_CH25 = TabulatedEmbeddedCubature{Rectangle}(;
+    description="Cools-Haegemans with 25 nodes (SQUARE_CH25)",
+    reference="https://link.springer.com/article/10.1007/BF01389339",
+    precision=35,
+    nodes=[
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "9.530899230491018626945996533246312660e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "9.530899230491018626945996533246312660e-01",
+        ],
+        [
+            "4.691007695089813730540034667536873396e-02",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "4.691007695089813730540034667536873396e-02",
+        ],
+        [
+            "7.692346551549900625567506592830037596e-01",
+            "7.692346551549900625567506592830037596e-01",
+        ],
+        [
+            "2.307653448450099374432493407169962404e-01",
+            "7.692346551549900625567506592830037596e-01",
+        ],
+        [
+            "7.692346551549900625567506592830037596e-01",
+            "2.307653448450099374432493407169962404e-01",
+        ],
+        [
+            "2.307653448450099374432493407169962404e-01",
+            "2.307653448450099374432493407169962404e-01",
+        ],
+        [
+            "9.530899229853897579996255404928620237e-01",
+            "9.530899229853897579996255404928620237e-01",
+        ],
+        [
+            "4.691007701461024200037445950713797632e-02",
+            "9.530899229853897579996255404928620237e-01",
+        ],
+        [
+            "9.530899229853897579996255404928620237e-01",
+            "4.691007701461024200037445950713797632e-02",
+        ],
+        [
+            "4.691007701461024200037445950713797632e-02",
+            "4.691007701461024200037445950713797632e-02",
+        ],
+        [
+            "9.530899229139767442924391519556961148e-01",
+            "7.692346549247062656221145758591156811e-01",
+        ],
+        [
+            "7.692346549247062656221145758591156811e-01",
+            "9.530899229139767442924391519556961148e-01",
+        ],
+        [
+            "4.691007708602325570756084804430388521e-02",
+            "7.692346549247062656221145758591156811e-01",
+        ],
+        [
+            "7.692346549247062656221145758591156811e-01",
+            "4.691007708602325570756084804430388521e-02",
+        ],
+        [
+            "9.530899229139767442924391519556961148e-01",
+            "2.307653450752937343778854241408843189e-01",
+        ],
+        [
+            "2.307653450752937343778854241408843189e-01",
+            "9.530899229139767442924391519556961148e-01",
+        ],
+        [
+            "4.691007708602325570756084804430388521e-02",
+            "2.307653450752937343778854241408843189e-01",
+        ],
+        [
+            "2.307653450752937343778854241408843189e-01",
+            "4.691007708602325570756084804430388521e-02",
+        ],
+        [
+            "7.692346549876881686600646748390376768e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "7.692346549876881686600646748390376768e-01",
+        ],
+        [
+            "2.307653450123118313399353251609623232e-01",
+            "5.000000000000000000000000000000000000e-01",
+        ],
+        [
+            "5.000000000000000000000000000000000000e-01",
+            "2.307653450123118313399353251609623232e-01",
+        ],
+    ],
+    weights_high=[
+        "8.090864180310654470524179109436875230e-02",
+        "3.369626803446566471100225222749401905e-02",
+        "3.369626803446566471100225222749401905e-02",
+        "3.369626803446566471100225222749401905e-02",
+        "3.369626803446566471100225222749401905e-02",
+        "5.727135095586947958087052565817859454e-02",
+        "5.727135095586947958087052565817859454e-02",
+        "5.727135095586947958087052565817859454e-02",
+        "5.727135095586947958087052565817859454e-02",
+        "1.403358722094394462690571699238021305e-02",
+        "1.403358722094394462690571699238021305e-02",
+        "1.403358722094394462690571699238021305e-02",
+        "1.403358722094394462690571699238021305e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "2.835000002587049491397105606308867037e-02",
+        "6.807163328620328507696894522217764455e-02",
+        "6.807163328620328507696894522217764455e-02",
+        "6.807163328620328507696894522217764455e-02",
+        "6.807163328620328507696894522217764455e-02",
+    ],
+    order_high=9,
+    weights_low=[
+        "1.689802306319903682683290932522631273e-01",
+        "5.773210696822923076352990074033502715e-02",
+        "5.773210696822923076352990074033502715e-02",
+        "5.773210696822923076352990074033502715e-02",
+        "5.773210696822923076352990074033502715e-02",
+        "1.098847683946998119578127419601697553e-01",
+        "1.098847683946998119578127419601697553e-01",
+        "1.098847683946998119578127419601697553e-01",
+        "1.098847683946998119578127419601697553e-01",
+        "2.059326850521098279376569809479221704e-02",
+        "2.059326850521098279376569809479221704e-02",
+        "2.059326850521098279376569809479221704e-02",
+        "2.059326850521098279376569809479221704e-02",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+        "9.772399236931191208904692945818609348e-03",
+    ],
+    order_low=7,
+)
